Twelve Tone Composition Tools
This tool was created to aid the composition of Twelve Tone style music. 

The user only needs to enter their prime row and press the build button to populate the whole Twelve Tone matrix. 
By using the other buttons on the right hand side of the the representation of the matrix can change from numerical to alphanumeric, flat musical notation, and sharp musical notation.